In newly decoupled code: added a doc string to inner class Price
I try to keep my code self-documenting but more importantly I want the code to be easily understood on first reading. My concern is that the control flow dictates the `Price` inner class has to be written precisely where it is but without knowing its purpose it could easily distract from the rest of ProductModel if it's not clear why it is there.