Update decorator.py
I would extend the decorator from the base class. The reason is that we still expect an italic bold text to be a text! So we expect the behaviours of the basic TextTag class to exists in the class that is being decorated. Example: If you add the following method to the  TextTag class, you expect your decorated objects still be able to call that method:
```
def say_hello():
    print 'hello I am text!'
```